refactor(v010): carve voicegw init out of _legacy
First per-command carve-out in section 2.

Adds voicegateway/cli/init.py owning the 'init' Typer command plus
the three helpers it alone uses (_PACKAGE_ROOT,
_EXAMPLE_CONFIG_CANDIDATES, _find_example_config). The package's
__init__.py now imports voicegateway.cli.init for its side effect:
the @app.command() decorator runs at import time and registers
'init' on the shared Typer app exactly as before. _legacy.py loses
~30 lines (the four moved members).

Bug fix bundled in: iteration 6's cli.py -> cli/_legacy.py rename
silently broke _PACKAGE_ROOT. The constant
'Path(__file__).parent.parent' used to resolve to the repo root
when the file was at voicegateway/cli.py; after the rename it
resolved to voicegateway/ (one level too deep), so
_find_example_config() always returned None and 'voicegw init'
silently fell back to the stub template instead of copying
voicegw.example.yaml. Tests passed because the fallback path also
creates the file. The carved init.py uses
Path(__file__).resolve().parent.parent.parent (one extra .parent)
which correctly lands on the repo root from the new location.
Verified: _find_example_config() now returns the real example file.

During the migration period submodules import 'app' and 'console'
directly from _legacy.py rather than from the package, because
ruff's isort rule reorders __init__.py imports in a way that
makes the package-namespace approach circular. When _legacy.py is
finally emptied, app and console move to __init__.py and every
submodule's import line flips in one final commit. Documented
inline in init.py.

All 13 commands still register on the shared app. pytest
tests/test_cli.py 46/46 pass; full suite 882 passed + 4 skipped;
ruff + mypy clean.
